Code for the assignment1
This version contains:
- More than 1 hidden layer
- ReLU activation function
- Leaky ReLU activation function

Most of the code was copied from the tutorial 2. I put in some places comments about what should be there. Softmax and Cross Entropy loss are needed only for the final output layer and that one propagation.
I will implement Adam another day as I already have code for that.